pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sentiment_allocine"
version = "0.1.0"
description = "French sentiment analysis on Allociné reviews - MLOps pipeline"
authors = [{name = "Votre Nom", email = "votre.email@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "transformers>=4.30",
    "torch>=2.0",
    "datasets>=2.14",
    "fastapi>=0.100",
    "uvicorn>=0.23",
    "streamlit>=1.25",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "loguru>=0.7",
    "pytest>=7.4",
    "pytest-cov>=4.1"
]

[project.optional-dependencies]
dev = ["black", "isort", "flake8", "pre-commit"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["src*"]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
addopts = "-ra -q --cov=src --cov-report=term-missing"
